+++ components/chrome/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import SocialLinks from './SocialLinks';
+import SocialLinks from '../SocialLinks';
 
 export default function Footer() {
